{"types":"C","runs":7}

+++ StringsExtra2.java
@@ -6,7 +6,7 @@ import acm.program.ConsoleProgram;
 public class StringsExtra2 extends ConsoleProgram {
 	public void run(){
 		String romanNum = readLine("Give me a Roman Number: ");
-		int number = 0;
+		double number = 0;
 		for(int i = 0; i < romanNum.length(); i++){
 			char charecter = romanNum.charAt(i);
 			if(charecter == 'I'){
